requirements.txt
dispy
pssh
paramiko
python-libnmap
scapy-python3
ansible
jupyter
psutil
